+++ module/comment/src/Infrastructure/Persistence/Projector/CommentContentChangedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Comment\Persistence\Dbal\Projector;
+namespace Ergonode\Comment\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Types;
